project-04_ai-code-review-agent/requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
PyGithub
python-dotenv
requests